+++ test/exportFile.js
@@ -13,19 +13,37 @@ test('Transform data when geojson already exists', t => {
   const options = {
     id: 'f445febc447d4cb696e71ea7816d65d5',
     layer: 0,
-    key: 'test',
-    format: 'csv',
-    name: 'test',
-    path: './test/output'
+    output: 'files/f445febc447d4cb696e71ea7816d65d5_0/full_0/test.csv',
+    source: `files/f445febc447d4cb696e71ea7816d65d5_0/full_0/test.geojson`
   }
   exportFile(options, err => {
     t.error(err, 'No error')
     const rows = []
-    _(fs.createReadStream(fileLoc(options)))
+    _(fs.createReadStream(fileLoc(options.output)))
     .split()
     .compact()
     .each(r => rows.push(r))
     .done(() => t.equal(rows.length, 101), 'CSV has all expected features')
+  })
+})
+
+test('Transform data with a filter when geojson already exists', t => {
+  t.plan(2)
+  const options = {
+    id: 'f445febc447d4cb696e71ea7816d65d5',
+    layer: 0,
+    where: 'FID > 50',
+    output: 'files/f445febc447d4cb696e71ea7816d65d5_0/test_filter/test.csv',
+    source: `files/f445febc447d4cb696e71ea7816d65d5_0/full_0/test.geojson`
+  }
+  exportFile(options, err => {
+    t.error(err, 'No error')
+    const rows = []
+    _(fs.createReadStream(fileLoc(options.output)))
+    .split()
+    .compact()
+    .each(r => rows.push(r))
+    .done(() => t.equal(rows.length, 51), 'CSV has all expected features')
   })
 })
 
@@ -44,17 +62,13 @@ test('Transform data when geojson does not exist', t => {
   const options = {
     id: 'test',
     layer: 0,
-    key: 'test',
-    format: 'csv',
-    name: 'test',
-    path: './test/output'
+    output: `files/test_0/full_0/test.csv`
   }
   exportFile(options, err => {
     t.error(err, 'No error')
 
-    const path = fileLoc(options)
     const rows = []
-    _(fs.createReadStream(path))
+    _(fs.createReadStream(fileLoc(options.output)))
     .split()
     .compact()
     .each(r => rows.push(r))
@@ -67,14 +81,11 @@ test('Transform data directly into geojson', t => {
   const options = {
     id: 'test',
     layer: 1,
-    key: 'test',
-    format: 'geojson',
-    name: 'test',
-    path: './test/output'
+    output: `files/test_0/full_1/test.geojson`
   }
   exportFile(options, err => {
     t.error(err, 'No error')
-    const path = fileLoc(options)
+    const path = fileLoc(options.output)
     const geojson = fs.readFileSync(path)
     const features = JSON.parse(geojson).features
     t.equal(features.length, 100, 'GeoJSON has all expected features')
@@ -86,27 +97,29 @@ test('Transform data from legacy geojson', t => {
   const options = {
     id: 'test',
     layer: 1,
-    key: 'test',
-    format: 'geojson',
-    name: 'test',
-    path: './test/output'
+    output: `files/test_0/full_1/test.csv`
   }
   exportFile(options, err => {
     t.error(err, 'No error')
-    const path = fileLoc(options)
-    const geojson = fs.readFileSync(path)
-    const features = JSON.parse(geojson).features
-    t.equal(features.length, 100, 'GeoJSON has all expected features')
+    const rows = []
+    _(fs.createReadStream(fileLoc(options.output)))
+    .split()
+    .compact()
+    .each(r => rows.push(r))
+    .done(() => t.equal(rows.length, 101), 'CSV has all expected features')
   })
 })
 
 test('Teardown', t => {
   rimraf.sync('./test/output')
-  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_0/test_0')
-  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_0/test_1')
+  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_0/full_0/test.csv')
+  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_0/full_1/test.csv')
+  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_3/full_3/test.csv')
+  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_0/full_1')
+  rimraf.sync('./test/data/files/f445febc447d4cb696e71ea7816d65d5_0/test_filter/')
   t.end()
 })
 
-function fileLoc (options) {
-  return `./test/data/files/${options.id}_${options.layer}/${options.key}/${options.name}.${options.format}`
+function fileLoc (output) {
+  return `./test/data/${output}`
 }